module_2/requirement.txt
Flask>=2.0.0
opencv-python>=4.5.0
ultralytics>=8.0.0
reportlab>=3.5.0
torch>=1.8.0
pyyaml>=5.4.0